pyproject.toml
[project]
name = "critter-capture"
version = "0.1.0"
description = "Animal species classification pipeline"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.0",
    "matplotlib>=3.7.1",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pillow>=9.5.0",
    "pyyaml>=6.0.1",
    "requests>=2.31.0",
    "scikit-learn>=1.3.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "tqdm>=4.65.0",
    "mlflow>=2.8.0",
    "ray[tune]>=2.7.0",
    "pydantic>=2.3.0",
    "iterative-stratification>=0.1.7",
    "boto3>=1.28.0",
    "streamlit>=1.30.0",
    "aiohttp>=3.9.0",
    "httpx>=0.25.0",
    "uvicorn>=0.23.0",
    "pyyaml-env-tag>=0.1",
    "python-dotenv>=1.0.0",
    "orjson>=3.9.0",
]


[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["critter_capture*"]
where = ["src"]

[tool.setuptools]
package-dir = {"" = "src"}
